migrate tests for cursor from PG to TCK
